+++ src/data-flow/service-registration.js
@@ -4,6 +4,7 @@ import { CommandFlowDescriptor } from './commands/command-flow-descriptor.js';
 import { EventFlowDescriptor } from './events/event-flow-descriptor.js';
 import { ExecutionOverviewDescriptor } from './execution/execution-overview-descriptor.js';
 import { ServiceIntegrationDescriptor } from './integrations/service-integration-descriptor.js';
+import { MessagingPatternsDescriptor } from './messaging/messaging-patterns-descriptor.js';
 import { QueryFlowDescriptor } from './queries/query-flow-descriptor.js';
 import { WorkflowOrchestrationDescriptor } from './workflows/workflow-orchestration-descriptor.js';
 
@@ -16,6 +17,7 @@ export function addDataFlow(services) {
   services.registerSingleton('CommandFlowDescriptor', () => new CommandFlowDescriptor());
   services.registerSingleton('EventFlowDescriptor', () => new EventFlowDescriptor());
   services.registerSingleton('ExecutionOverviewDescriptor', () => new ExecutionOverviewDescriptor());
+  services.registerSingleton('MessagingPatternsDescriptor', () => new MessagingPatternsDescriptor());
   services.registerSingleton('QueryFlowDescriptor', () => new QueryFlowDescriptor());
   services.registerSingleton('ServiceIntegrationDescriptor', () => new ServiceIntegrationDescriptor());
   services.registerSingleton('WorkflowOrchestrationDescriptor', () => new WorkflowOrchestrationDescriptor());